design:리뷰상세페이지 반응형 적용
